+++ orchestrator/relayer/src/main_loop.rs
@@ -1,4 +1,7 @@
-use crate::{batch_relaying::relay_batches, valset_relaying::relay_valsets};
+use crate::{
+    batch_relaying::relay_batches, find_latest_valset::find_latest_valset,
+    valset_relaying::relay_valsets,
+};
 use clarity::address::Address as EthAddress;
 use clarity::PrivateKey as EthPrivateKey;
 use peggy_proto::peggy::query_client::QueryClient as PeggyQueryClient;
@@ -20,7 +23,23 @@ pub async fn relayer_main_loop(
     let mut grpc_client = grpc_client;
     loop {
         let loop_start = Instant::now();
+
+        let our_ethereum_address = ethereum_key.to_public_key().unwrap();
+        let current_valset = find_latest_valset(
+            &mut grpc_client,
+            our_ethereum_address,
+            peggy_contract_address,
+            &web3,
+        )
+        .await;
+        if current_valset.is_err() {
+            error!("Could not get current valset!");
+            continue;
+        }
+        let current_valset = current_valset.unwrap();
+
         relay_valsets(
+            current_valset.clone(),
             ethereum_key,
             &web3,
             &mut grpc_client,
@@ -30,6 +49,7 @@ pub async fn relayer_main_loop(
         .await;
 
         relay_batches(
+            current_valset,
             ethereum_key,
             &web3,
             &mut grpc_client,
